# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amane API",
    "version": "0.7.2"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "description": "默认只返回链根 (roots_only). 子任务由 /children 按需加载;\n显式 root_task_id 时返回该链全部任务.",
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskStatus"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by task status (multiple)",
              "title": "Status"
            },
            "description": "Filter by task status (multiple)"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskType"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by task type (multiple)",
              "title": "Type"
            },
            "description": "Filter by task type (multiple)"
          },
          {
            "name": "root_task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a task chain (all tasks sharing the root)",
              "title": "Root Task Id"
            },
            "description": "Filter to a task chain (all tasks sharing the root)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TaskSortField",
              "description": "Sort field",
              "default": "created_at"
            },
            "description": "Sort field"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "Sort order",
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Submit Task",
        "operationId": "submit_task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/OrganizeSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/ScrapeSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/CleanupSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/UpscaleSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/R18ImportSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/ActorScrapeSubmission"
                  },
                  {
                    "$ref": "#/components/schemas/RescrapeSubmission"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/children": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Children",
        "description": "直接后继 (TaskLink 出边). total 为出边总数, 不受本页截断.",
        "operationId": "get_task_children",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskChildListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/batch": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Batch Tasks",
        "description": "``task_ids`` 与 status/type 互斥.",
        "operationId": "batch_tasks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/report": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Report",
        "description": "面向 UI 的投影, 非完整记录导出. 仅终态可用.",
        "operationId": "get_task_report",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/record": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Record",
        "description": "导出任务记录 (zip). 默认脱敏; include_secrets=true 需本地存在 .secrets.hot.json.",
        "operationId": "get_task_record",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "include_secrets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include plaintext cookies/tokens from local secrets snapshot",
              "default": false,
              "title": "Include Secrets"
            },
            "description": "Include plaintext cookies/tokens from local secrets snapshot"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActorScrapeSubmission": {
        "properties": {
          "type": {
            "type": "string",
            "const": "actor_scrape",
            "title": "Type"
          },
          "actor_id": {
            "type": "integer",
            "title": "Actor Id",
            "description": "Actor 实体 ID"
          },
          "use_cache": {
            "items": {
              "$ref": "#/components/schemas/CacheKind"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Use Cache",
            "description": "启用的缓存种类 (metadata: 复用 Actor.raw; trans: 预留译文). 空集 = 全部强制刷新"
          }
        },
        "type": "object",
        "required": [
          "type",
          "actor_id"
        ],
        "title": "ActorScrapeSubmission"
      },
...
      "CacheKind": {
        "type": "string",
        "enum": [
          "metadata",
          "trans"
        ],
        "title": "CacheKind",
        "description": "刮削可复用的缓存种类. use_cache 为其集合: 含某项 = 该缓存生效, 不含 = 强制刷新该项."
      },
      "CleanupSubmission": {
        "properties": {
          "remove_missing_files": {
            "type": "boolean",
            "title": "Remove Missing Files",
            "description": "删除磁盘上不存在的 MediaFile 记录 (不触碰 Metadata)",
            "default": true
          },
          "remove_unreferenced_resources": {
            "type": "boolean",
            "title": "Remove Unreferenced Resources",
            "description": "删除不被任何 Metadata URL 字段引用的 Resource (含派生裁剪)",
            "default": true
          },
          "type": {
            "type": "string",
            "const": "cleanup",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "CleanupSubmission"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaFileStatus": {
        "type": "string",
        "enum": [
          "pending",
          "scraped",
          "failed",
          "skip"
        ],
        "title": "MediaFileStatus"
      },
...
      "OrganizeSubmission": {
        "properties": {
          "library_id": {
            "type": "integer",
            "title": "Library Id",
            "description": "所属 Library ID; 扫描/整理在该媒体库下进行",
            "x-widget": "LibraryPicker"
          },
          "recursive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recursive",
            "description": "覆盖 Library 的 recursive; None 沿用库设置"
          },
          "patterns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patterns",
            "description": "覆盖 Library 的 patterns; None 沿用库设置"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "要扫描的目录路径 (覆盖 Library 路径, 必须为 Library 子目录).",
            "default": "",
            "x-path-type": "directory",
            "x-widget": "PathPicker"
          },
          "write_nfo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Write Nfo",
            "description": "覆盖 Library.write_nfo; None 沿用库设置"
          },
          "copy_resources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DownloadableResource"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Copy Resources",
            "description": "覆盖 Library.copy_resources; None 沿用库设置"
          },
          "type": {
            "type": "string",
            "const": "organize",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "type"
        ],
        "title": "OrganizeSubmission"
      },
...
      "R18ImportSubmission": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "type": {
            "type": "string",
            "const": "r18_import",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "R18ImportSubmission"
      },
      "RefreshSubmission": {
        "properties": {
          "library_id": {
            "type": "integer",
            "title": "Library Id",
            "description": "所属 Library ID; 扫描/整理在该媒体库下进行",
            "x-widget": "LibraryPicker"
          },
          "recursive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recursive",
            "description": "覆盖 Library 的 recursive; None 沿用库设置"
          },
          "patterns": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patterns",
            "description": "覆盖 Library 的 patterns; None 沿用库设置"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "要扫描的目录路径 (覆盖 Library 路径, 必须为 Library 子目录).",
            "default": "",
            "x-path-type": "directory",
            "x-widget": "PathPicker"
          },
          "scan": {
            "items": {
              "$ref": "#/components/schemas/ScanMode"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Scan",
            "default": [
              "add"
            ]
          },
          "scrape": {
            "items": {
              "$ref": "#/components/schemas/MediaFileStatus"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Scrape",
            "default": [
              "pending"
            ]
          },
          "use_cache": {
            "items": {
              "$ref": "#/components/schemas/CacheKind"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Use Cache",
            "default": [
              "metadata",
              "trans"
            ]
          },
          "type": {
            "type": "string",
            "const": "refresh",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "type"
        ],
        "title": "RefreshSubmission"
      },
...
      "RescrapeSubmission": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "单次最多补刮的元数据数 (避免长占 worker 队列)",
            "default": 100
          },
          "min_age_days": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Age Days",
            "description": "仅补刮 updated_at 距今超过该天数的条目; None 不设门槛"
          },
          "type": {
            "type": "string",
            "const": "rescrape",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "RescrapeSubmission"
      },
...
      "ScanMode": {
        "type": "string",
        "enum": [
          "add",
          "remove"
        ],
        "title": "ScanMode"
      },
...
      "ScrapeSubmission": {
        "properties": {
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "media_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Id",
            "description": "MediaFile ID. 通常仅用于内部提交任务，手动提交无需指定"
          },
          "content_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentType"
              },
              {
                "type": "null"
              }
            ],
            "description": "内容类型; 未给出时: 仅 media_id 按文件路径推断, 有 number 时按番号推断"
          },
          "use_cache": {
            "items": {
              "$ref": "#/components/schemas/CacheKind"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Use Cache",
            "description": "启用的缓存种类 (metadata: 复用 DB per-site 快照; trans: 复用译文). 空集 = 全部强制刷新"
          },
          "type": {
            "type": "string",
            "const": "scrape",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "ScrapeSubmission"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
...
      "TaskBatchRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TaskBatchAction"
          },
          "task_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Ids"
          },
          "status": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "type": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TaskType"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "TaskBatchRequest"
      },
...
      "TaskChildListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskChildResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TaskChildListResponse"
      },
...
      "TaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TaskListResponse"
      },
...
      "TaskSortField": {
        "type": "string",
        "enum": [
          "type",
          "status",
          "priority",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "TaskSortField"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "done",
          "failed"
        ],
        "title": "TaskStatus"
      },
      "TaskType": {
        "type": "string",
        "enum": [
          "scrape",
          "organize",
          "refresh",
          "cleanup",
          "upscale",
          "r18_import",
          "actor_scrape",
          "rescrape"
        ],
        "title": "TaskType"
      },
...
      "UpscaleSubmission": {
        "properties": {
          "max_dim_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Dim Threshold"
          },
          "max_bytes_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Bytes Threshold"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 200
          },
          "type": {
            "type": "string",
            "const": "upscale",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "UpscaleSubmission"
      },
...
    }
  }
}